requirements.txt
fredapi
yfinance
xgboost
scikit-learn
anthropic
streamlit
pandas
plotly
python-dotenv